refactor: 💡 move VenueLink into a separate component
