+++ src/components/shared/eventcard/eventcard.js
@@ -11,6 +11,8 @@ import EventPlaceholderImage from 'images/event-placeholder.png';
 import ClockIcon from 'images/clock.svg';
 import MapMarkerIcon from 'images/map-marker.svg';
 
+import VenueLink from 'components/shared/venuelink';
+
 import {
   Article,
   ImageWrapper,
@@ -19,7 +21,6 @@ import {
   IconList,
   IconRow,
   IconText,
-  IconLink,
 } from './eventcard.css';
 
 import { parse, format } from 'date-fns';
@@ -29,31 +30,6 @@ const EventCard = ({ event }) => {
 
   const parsedDateTime = parse(time, 'T', new Date());
   const formattedTime = format(parsedDateTime, 'h:mm a');
-
-  const venueLink = () => {
-    if (venue) {
-      const encodedAddress = encodeURI(
-        `${venue.name}, ${venue.address_1}, ${venue.city}, ${venue.state}`
-      );
-      const directionsUrl = `https://www.google.com/maps/search/?api=1&query=${encodedAddress}`;
-      return (
-        <IconRow mt={3}>
-          <Box flex="none">
-            <MapMarkerIcon />
-          </Box>
-          <IconLink
-            as="a"
-            href={directionsUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            {venue.name}
-          </IconLink>
-        </IconRow>
-      );
-    }
-    return null;
-  };
 
   return (
     <Article>
@@ -87,7 +63,14 @@ const EventCard = ({ event }) => {
             </Box>
             <IconText>{formattedTime}</IconText>
           </IconRow>
-          {venueLink()}
+          {venue && (
+            <IconRow mt={3}>
+              <Box flex="none">
+                <MapMarkerIcon />
+              </Box>
+              <VenueLink venue={venue} />
+            </IconRow>
+          )}
         </IconList>
         <Box mt={{ _: 4, sm: 0 }}>
           <Button
